backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-bcrypt==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-cors==4.0.0
PyMuPDF==1.23.8
python-docx==0.8.11
docx2txt==0.8
google-genai==1.16.1